arm64: dts: qcom: msm8909/16/39: Add standalone RPM node
Rather than having the RPM SMD channels as the only child of a dummy
SMD node, switch to representing the RPM as remoteproc like all the
other remoteprocs (WCNSS, modem DSP).

This allows assigning additional resources to it, like a reset control
or an IOMMU which is necessary when using the TF-A TZ firmware for
MSM8909/MSM8916/MSM8939.

Signed-off-by: Stephan Gerhold <stephan@gerhold.net>