arm64: dts: qcom: msm8916-alcatel-idol347: Fix mount-matrix for Android
Not sure if the current mount-matrix is wrong, or if the coordinate
system is different on Android but it causes wrong screen rotation.

Change the mount-matrix for all sensors to match the one seemingly used
on the downstream kernel, which should hopefully be correct for Android.
It fixes the screen rotation at least.

	kionix,accel_direction = <8>;
	akm,layout = <0x1>;
	bmg,place = <6>;

Signed-off-by: Stephan Gerhold <stephan@gerhold.net>