dt-bindings: power: rpmpd: Add MSM8917, MSM8937 and QM215
The MSM8917, MSM8937 and QM215 SoCs have VDDCX and VDDMD power domains
controlled in voltage level mode. Define the MSM8937 and QM215 power
domains as aliases because these SoCs are similar to MSM8917 and may
share some parts of the device tree.

Also add the compatibles for these SoCs to the documentation.

Signed-off-by: Otto Pflüger <otto.pflueger@abscue.de>